@@ -1,12 +1,15 @@
-const { Router } = require("express");
+// construtor das minhas rotas, tenho mais controles sobre elas
+// separando em um arquivo "routes.js"
+
+import { Router } from "express";
+import customers from "./app/controllers/CustomersController";
+
 const routes = new Router();
 
-const customers = require("./app/controllers/CustomersController");
+routes.get("/customers", customers.index);
+routes.get("/customers/:id", customers.show);
+routes.post("/customers", customers.create);
+routes.put("/customers/:id", customers.update);
+routes.delete("/customers/:id", customers.destroy);
 
-routes.get("/customers", customers.index)
-routes.get("/customers/:id", customers.show)
-routes.post("/customers", customers.create)
-routes.put("/customers/:id", customers.update)
-routes.delete("/customers/:id", customers.destroy)
-
-module.exports = routes;
+export default routes;
