@@ -1,4 +1,5 @@
-const app = require("./app");
-app.listen(3000, ()=>{
-    console.log("server em: http://localhost:3000/"); 
-})
+import app from "./app";
+
+app.listen(3000, () => {
+  //console.log("server em: http://localhost:3000/");
+});
